% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme.gev.R
\name{init.glme}
\alias{init.glme}
\title{Initialize parameters for Generalized L-moments estimation of GEV distribution}
\usage{
init.glme(xdat, ntry = ntry)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted.}

\item{ntry}{Number of initial parameter sets to generate.}
}
\value{
A matrix with `ntry` rows and 3 columns, where each row represents
a set of initial parameters (location, scale, shape) for the GEV distribution.
}
\description{
This function initializes parameters for the Generalized L-moments estimation
of the Generalized Extreme Value (GEV) distribution.
}
\details{
The function generates `ntry` sets of initial parameters for the GEV distribution.
It uses L-moment estimates as a starting point and then generates additional
sets of parameters using random perturbations. This approach increases the
likelihood of finding a global optimum in the subsequent optimization process.

The function ensures that:
1. The scale parameter (sigma) is always positive.
2. The shape parameter (xi) is constrained between -0.7 and 0.7.
}
\author{
[Jeong-Soo Park]
}
